+++ packages/gov-ui-kit/CHANGELOG.md
@@ -11,6 +11,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 
 -   Update `ProposalVoting.Container` module component to make stages accordion controlled and support updating the
     current active stage programmatically
+-   Fix `VoteProposalDataListItem` module component to truncate long proposal IDs
 
 ### Changed